+++ include/mu/tiny/test/Installer.hpp
@@ -43,7 +43,7 @@ public:
       const char* group_name,
       const char* test_name,
       const char* file_name,
-      size_t line_number
+      int line_number
   ) noexcept;
   virtual ~Installer() = default;
   Installer(const Installer&) = delete;
