+++ src/test/Registry.cpp
@@ -127,8 +127,7 @@ void Registry::list_test_locations(Result& result)
     test_location += ".";
     test_location += test->get_file();
     test_location += ".";
-    test_location +=
-        string_from_format("%d\n", static_cast<int>(test->get_line_number()));
+    test_location += string_from_format("%d\n", test->get_line_number());
 
     test_locations += test_location;
   }
@@ -151,8 +150,7 @@ void Registry::list_ordered_test_locations(Result& result)
     test_location += ".";
     test_location += test->get_file();
     test_location += ".";
-    test_location +=
-        string_from_format("%d\n", static_cast<int>(test->get_line_number()));
+    test_location += string_from_format("%d\n", test->get_line_number());
 
     test_locations += test_location;
   }
@@ -171,7 +169,7 @@ void Registry::list_test_group_locations(Result& result)
     entry += ".";
     entry += g->get_file();
     entry += ".";
-    entry += string_from_format("%d\n", static_cast<int>(g->get_line_number()));
+    entry += string_from_format("%d\n", g->get_line_number());
 
     group_locations += entry;
   }
